pyproject.toml
[project]
name = "weather-station"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "environs>=15.0.1",
    "nicegui>=3.12.0",
    "rpi-bme280>=0.2.4",
]
